+++ packages/store/demo/benchmark.mjs
@@ -1,13 +1,13 @@
 import {createLogger} from '@alwatr/logger';
-import {AlwatrStore, Region} from '@alwatr/nitrobase';
+import {AlwatrNitrobase, Region} from '@alwatr/nitrobase';
 import {waitForTimeout} from '@alwatr/wait';
 
 (async function () {
-  const logger = createLogger('AlwatrStore/Demo', true);
-  logger.banner('AlwatrStore/Demo');
+  const logger = createLogger('AlwatrNitrobase/Demo', true);
+  logger.banner('AlwatrNitrobase/Demo');
 
   // Create a new storage instance
-  const alwatrStore = new AlwatrStore({
+  const alwatrStore = new AlwatrNitrobase({
     rootPath: './db',
     defaultChangeDebounce: 1000, // for test
   });